+++ api/src/org/labkey/api/assay/AssayService.java
@@ -134,7 +134,7 @@ public interface AssayService
     void indexAssay(SearchService.IndexTask task, Container c, ExpProtocol protocol);
     void indexAssays(SearchService.IndexTask task, Container c);
 
-    void indexAssayRun(int expRunRowId);
+    void indexAssayRun(long expRunRowId);
 
     void deindexAssays(@NotNull Collection<? extends ExpProtocol> expProtocols);
 
@@ -222,7 +222,7 @@ public interface AssayService
      *
      * @return the count of flags deleted
      */
-    int deleteFlagsForRun(Container container, User user, AssayProvider provider, int runId);
+    int deleteFlagsForRun(Container container, User user, AssayProvider provider, long runId);
 
     /**
      * Delete the specified flag.
@@ -234,7 +234,7 @@ public interface AssayService
      * Returns the flags for the specified run.
      * The assay provider must implement an instance of the AssayFlagHandler interface
      */
-    <FlagType extends ExpQCFlag> List<FlagType> getFlags(AssayProvider provider, int runId, Class<FlagType> cls);
+    <FlagType extends ExpQCFlag> List<FlagType> getFlags(AssayProvider provider, long runId, Class<FlagType> cls);
 
     void onBeforeAssayResultDelete(Container container, User user, ExpRun run, Map<String, Object> resultRow);
